Refactor modules state api objects (#537)
* Remove objets.rs and add everything in mod.rs

* Remove state_api_objets.rs and add everything in mod.rs

* import only used structs

* fix conflicts

---------
