+++ src/core/syscalls/business_logic_syscall_handler.rs
@@ -18,7 +18,7 @@ use super::{
     },
     syscall_response::{CallContractResponse, FailureReason, ResponseBody},
 };
-use crate::business_logic::state::state_api_objects::BlockInfo;
+use crate::business_logic::state::BlockInfo;
 use crate::business_logic::transaction::error::TransactionError;
 use crate::utils::calculate_sn_keccak;
 use crate::{
